+++ packages/framer-motion/src/context/MotionConfigContext.tsx
@@ -33,6 +33,15 @@ export interface MotionConfigContext {
      * @public
      */
     reducedMotion?: ReducedMotionConfig
+
+    /**
+     * A custom `nonce` attribute used when wanting to enforce a Content Security Policy (CSP).
+     * For more details see:
+     * https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/style-src#unsafe_inline_styles
+     *
+     * @public
+     */
+    nonce?: string
 }
 
 /**