+++ packages/framer-motion/src/render/dom/utils/is-css-variable.ts
@@ -4,7 +4,7 @@ export type CSSVariableToken = `var(${CSSVariableName})`
 
 const checkStringStartsWith =
     <T extends string>(token: string) =>
-    (key?: string): key is T =>
+    (key?: string | number | null): key is T =>
         typeof key === "string" && key.startsWith(token)
 
 export const isCSSVariableName = checkStringStartsWith<CSSVariableName>("--")
@@ -22,4 +22,4 @@ export const isCSSVariableToken = (
 }
 
 const singleCssVariableRegex =
-    /var\s*\(\s*--[\w-]+(\s*,\s*(?:(?:[^)(]|\((?:[^)(]+|\([^)(]*\))*\))*)+)?\s*\)$/i
+    /var\(--(?:[\w-]+\s*|[\w-]+\s*,(?:\s*[^)(\s]|\s*\((?:[^)(]|\([^)(]*\))*\))+\s*)\)$/iu